requirements.txt
certifi==2024.2.2
packaging==24.0
wheel==0.42.0
yt-dlp==2024.3.10
youtube-transcript-api==0.6.2
openai-whisper==20231117
ffmpeg-python==0.2.0
pytube==15.0.0
youtube-dl==2021.12.17
tqdm==4.66.2
selenium==4.18.1
undetected-chromedriver==3.5.5
python-dotenv==1.0.1
google-api-python-client==2.123.0
pandas==2.2.1
